fix(AddressBar): Various fixes to address bar buttons / actions
